requirements.txt
streamlit
trimesh
numpy
requests
Pillow
